source/requirements/base.txt
django
django-rest-auth
djangorestframework
djangorestframework-jwt
envparse
gunicorn
psycopg2-binary
pytz
